fix: mq separate new ordered message queue
